M 优化授权帐户ID为int64

+++ model/sql/advertising/advertising_media_auth.go
@@ -29,7 +29,7 @@ type DimAdvertisingMediaAuthModel struct {
 	sql2.SqlBaseModel
 	PlatformId            int64                 `json:"platform_id" gorm:"column:platform_id;default:0;comment:平台ID;uniqueIndex:ix_plat_account"`
 	AccountName           string                `json:"account_name" gorm:"size:100;column:account_name;default:'';comment:帐户名称"`
-	AccountId             uint64                `json:"account_id" gorm:"size:100;column:account_id;default:0;comment:帐户ID;uniqueIndex:ix_plat_account"`
+	AccountId             int                   `json:"account_id" gorm:"size:100;column:account_id;default:0;comment:帐户ID;uniqueIndex:ix_plat_account"`
 	Status                string                `json:"status" gorm:"size:100;column:status;default:'';comment:状态"`
 	Extension             string                `json:"extension" gorm:"type:json;column:extension;comment:扩展字段"`
 	AccessToken           string                `json:"access_token" gorm:"size:100;column:access_token;default:'';comment:授权令牌"`
